requirements.txt
numpy
pandas
scipy
matplotlib
Pillow